requirements.txt
asgiref==3.6.0
cssselect2==0.7.0
Django==3.2.12
django-cors-headers==3.13.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-polymorphic==3.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djoser==2.0.3
easy-thumbnails==2.8.5
fuzzywuzzy==0.18.0
gunicorn==20.1.0
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.1
Pillow==9.4.0
psycopg2-binary==2.9.5
pytz==2022.7
reportlab==3.6.12
sqlparse==0.4.3
svglib==1.5.1
tinycss2==1.2.1
Unidecode==1.1.2
webencodings==0.5.1
